+++ spec/test_app/scripts/bad_syntax.sh
@@ -1,0 +1,9 @@
+# Tests that Bashcov::Detective excludes shell scripts with
+#   1. A shellscript extension like .sh or .bash *and*
+#   2. Contents that *do not* follow valid shell syntax
+
+confused = true
+
+if confused
+  puts 'I though this was supposed to be a shell script?'
+end