+++ spec/test_app/scripts/no_shebang.sh
@@ -1,0 +1,5 @@
+# Tests that Bashcov::Detective properly detects shell scripts with
+#   1. A shellscript extension like .sh or .bash *and*
+#   2. Contents that follow valid shell syntax
+
+echo "no shebang here!"